Bump version to 2.1.0 and refactor country fetching functions for improved structure and field selection
