+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@yusifaliyevpro/countries",
-  "version": "2.0.4",
+  "version": "2.1.0",
   "publishConfig": {
     "access": "public"
   },
